[Hopsworks 181] and  	[HOPSWORKS-95] (#636)
* [HOPSWORKS-181] fix Project Administration to comeback to what it was before to get brocken.

* [HOPSWORKS-95] change HDFS quota in Projects Management to be able to indicate the unit in which we enter it (MB, GB or TB)

+++ hopsworks-common/src/main/java/io/hops/hopsworks/common/dao/jobs/quota/YarnProjectsQuotaFacade.java
@@ -69,7 +69,7 @@ public class YarnProjectsQuotaFacade extends
     return query.getResultList();
   }
 
-  public void changeYarnQuota(String projectname, int quota) {
+  public void changeYarnQuota(String projectname, float quota) {
     YarnProjectsQuota project = findByProjectName(projectname);
     if (project != null) {
       project.setQuotaRemaining(quota);
